Create README.md
In this project, i use python as major and css as minor. It is such  authentic and resourceful project that describe detailed analyzation with the help of dash library as well as other library such as numpy, plotly and other. Here, flask is use as to create dashboard. Flask is designed to be easy to use and extend, and it has a large and active community of developers.